test(KEN-801): cut the test surface to the Done-when [no-changelog]
Overseer ruling on #2024. One control for the Done-when and nothing else this
branch wrote: an update started from the Projects tab, then the Overview's
files and version and the header's update offer all read as the state after
it. Its must-fail control is the commit dependency in usePackageData, which
turns it red.

Everything else this branch added to the tests goes: the ranking-clause
cases, the per-read settled and readOrder cases, the plugin, fork and derived
cases, the standing-note cases and the retry cases, with package-reads and
package-update deleted whole.

Moves come back rather than going with them, since they are coverage main
ships today: the version-actions block returns to use-package-data.test.ts
and package-version-actions.test.ts is gone, the note describe returns to
package.test.tsx, and the shared fixture file is gone because those fixtures
came out of that file in the first place. Two imports follow the production
module split, which does not change.

No production code changes.

The restored package.test.tsx is 874 lines against an 800-line class, 714 of
them main's own. The ruling says report rather than split a second time, so
it takes a baseline row: the file is back to what main carries plus the one
control the issue exists for.
